==> 3-NER_evaluation_v3-Copy2.ipynb <==
# https://www.datacamp.com/tutorial/fuzzy-string-python
# https://towardsdatascience.com/fuzzy-string-matching-in-python-68f240d910fe#:~:text=Fuzzywuzzy%20is%20a%20python%20library,showcase%20them%20on%20one%20platform.


# !pip install wandb
# !pip install transformers
# !pip install sentencepiece

from transformers import T5TokenizerFast, T5ForConditionalGeneration 
from transformers import Trainer

import os
import time 

os.environ["CUDA_VISIBLE_DEVICES"] = "1,2,3" 
import pandas as pd
import numpy as np
import torch
import torchvision
import Levenshtein
from fuzzywuzzy import fuzz
from datasets import list_metrics
from sklearn.metrics import precision_recall_fscore_support

device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")

# wandb.login()

# os.environ["WANDB_DISABLED"] = "true"

device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")

# Create torch dataset
# https://towardsdatascience.com/fine-tuning-pretrained-nlp-models-with-huggingfaces-trainer-6326a4456e7b

class Dataset(torch.utils.data.Dataset):
    def __init__(self, encodings, labels=None):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        if self.labels:
            item["labels"] = torch.tensor(self.labels['input_ids'][idx])
        return item

    def __len__(self):
        return len(self.encodings["input_ids"])

# conellpp_test.to_csv('./1-NER_Data/1-csv_format/test/conell_test_data.csv', index = False)
# wikineural_test.to_csv('./1-NER_Data/1-csv_format/test/wiki_neural_test_data.csv', index = False)
# onto_test.to_csv('./1-NER_Data/1-csv_format/test/onto_test_data.csv', index = False)
# wnut_test.to_csv('./1-NER_Data/1-csv_format/test/wnut_test_data.csv', index = False)

def read_test_data(path):
    return pd.read_csv(path)

def encode_input_as_tokens(dataset):
    tokens = []
    for text in dataset["input_text"]:
        tokens.append(tokenizer('ner: ' + text, return_tensors="pt", padding=True, truncation=True, max_length=512).input_ids)
    return tokens

def predict_ner(model, tokens):
    results = []
    model = model.to(device)
    counter =0
    for token in tokens:
        results.append(model.generate(token.to(device))) 
        counter +=1
        if counter % 100==0:
            print(counter)
    return results

def decode_prediction(predicted_tokens):
    final_ouput_2 = []
    for result in predicted_tokens:
        final_ouput_2.append(tokenizer.decode(result[0], skip_special_tokens=True))
    return final_ouput_2

def compare_prediction_with_target(predictions, target_text):
    counter = 0
    for i in range(len(predictions)):
        if predictions[i] != target_text[i]:
            counter += 1
    return 1- counter/len(predictions)

def compute_metrics(pred):
    labels = pred.label_ids
    preds = pred.predictions.argmax(-1)
    precision, recall, f1, _ = precision_recall_fscore_support(labels, preds, average='binary')
    acc = accuracy_score(labels, preds)
    return {
        'accuracy': acc,
        'f1': f1,
        'precision': precision,
        'recall': recall
    }

max_source_length = 512
max_target_length = 256

tokenizer = T5TokenizerFast.from_pretrained("t5-base")
model_path = "./NER_normal_diff_eval_steps_patience_5/checkpoint-37430"
model = T5ForConditionalGeneration.from_pretrained(model_path).to(device)

# from transformers import Seq2SeqTrainingArguments

# test_args = Seq2SeqTrainingArguments(
#     "test_trainer",
#     per_device_train_batch_size = 8,
#     per_device_eval_batch_size = 8,
#     eval_accumulation_steps = 50,  # VIP
#     predict_with_generate=True
# )

# from transformers import Seq2SeqTrainer

# trainer = Seq2SeqTrainer(
#     model=model, 
#     args=test_args
# #     train_dataset= train_dataset, 
# #     eval_dataset= eval_dataset
# )

## conell test data 

conell_test_data = read_test_data('./1-NER_Data/1-csv_format/test/conell_test_data.csv')
conell_test_data

conell_test_data['target_text'].str.len().hist()

conell_tokens = encode_input_as_tokens(conell_test_data)

len(conell_tokens)

start = time.time()
counter = 0
conell_pred = []
for token in conell_tokens:
    if counter%200 == 0:
        print(counter)
    counter += 1
    prediction = model.generate(token.to(device))
    conell_pred.append(prediction)
    
print(time.time() - start)

# conell_pred = predict_ner(model, conell_tokens)

conell_pred_decoded = decode_prediction(conell_pred)

conell_target_text = conell_test_data['target_text']

conell_accuracy = compare_prediction_with_target(conell_pred_decoded, conell_target_text)
conell_accuracy

conell_accuracy = compare_prediction_with_target(conell_pred_decoded, conell_target_text)
conell_accuracy

# lower_normal
conell_accuracy = compare_prediction_with_target(conell_pred_decoded, conell_target_text)
conell_accuracy

# predictions = pd.DataFrame({'predictions':conell_pred_decoded, 'actual':conell_target_text}) 

# predictions.to_csv('./1-NER_Data/2-predictions/lower_normal_conell.csv')

conell_false_pos = 0
conell_false_neg = 0
conell_true_pos = 0
conell_errors = 0

for i in range(len(conell_pred_decoded)):
    split_predicted_entities = conell_pred_decoded[i].split(',')
    split_target_entities = conell_target_text[i].split(',')
    
    for pred in split_predicted_entities:
        if pred in split_target_entities:
            conell_true_pos += 1
        else:
            print(split_predicted_entities) 
            print(split_target_entities)
            conell_errors += 1
        if len(split_predicted_entities) > len(split_target_entities):
            conell_false_pos += len(split_predicted_entities) - len(split_target_entities)
        elif len(split_predicted_entities) < len(split_target_entities):
            conell_false_neg += len(split_target_entities) - len(split_predicted_entities)

conell_errors

conell_true_pos/(conell_true_pos+conell_errors)

# Precision = TP/TP+FP
conell_true_pos/(conell_true_pos+conell_false_pos+conell_errors)

# Recall = TP/TP+FN
conell_true_pos/(conell_true_pos+conell_false_neg+conell_errors)

# conell_false_pos
# conell_false_neg
conell_errors

conell_true_pos/(conell_true_pos+conell_errors)

# Precision = TP/TP+FP
conell_true_pos/(conell_true_pos+conell_false_pos+conell_errors)

# Recall = TP/TP+FN
conell_true_pos/(conell_true_pos+conell_false_neg+conell_errors)

# lower_normal = 0.88
conell_true_pos/(conell_true_pos+conell_errors)

# lower_normal
# Precision = TP/TP+FP = 0.859
conell_true_pos/(conell_true_pos+conell_false_pos+conell_errors)

conell_true_pos/(conell_true_pos+conell_errors)

# Precision = TP/TP+FP
conell_true_pos/(conell_true_pos+conell_false_pos+conell_errors)

# Recall = TP/TP+FN
conell_true_pos/(conell_true_pos+conell_false_neg+conell_errors)

# lower_normal = 0.88
conell_true_pos/(conell_true_pos+conell_errors)

# lower_normal
# Precision = TP/TP+FP = 0.859
conell_true_pos/(conell_true_pos+conell_false_pos+conell_errors)

# lower_normal
# Recall = TP/TP+FN = 0.63
conell_true_pos/(conell_true_pos+conell_false_neg+conell_errors)

f1_score(conell_target_text, conell_pred_decoded, average='macro')

f1_score(conell_target_text, conell_pred_decoded, average='micro')

f1_score(conell_target_text, conell_pred_decoded, average='weighted')

target_encodings = []
for text in conell_test_data["target_text"]:
    target_encodings.append(tokenizer('ner: ' + text, return_tensors="pt", padding=True, truncation=True, max_length=512).input_ids)

len(conell_tokens)

len(target_encodings[0][0])

f1_score(target_encodings[0][0], conell_tokens[0][0])

score = 0
for i in range(len(conell_tokens)):
    score += f1_score(target_encodings[i], conell_tokens[i])
score/len(conell_tokens)

## wikineural test data

wiki_test_data = read_test_data('./1-NER_Data/1-csv_format/test/wiki_neural_test_data.csv')
wiki_test_data

wiki_tokens = encode_input_as_tokens(wiki_test_data)

len(wiki_tokens)

wiki_pred = predict_ner(model, wiki_tokens)

wiki_pred_decoded = decode_prediction(wiki_pred)

wiki_target_text = wiki_test_data['target_text']

wiki_accuracy = compare_prediction_with_target(wiki_pred_decoded, wiki_target_text)
wiki_accuracy

evaluations(predictions=conell_pred_decoded, references=conell_target_text):

## onto test data

onto_test_data = read_test_data('./1-NER_Data/1-csv_format/test/onto_test_data.csv')
onto_test_data

onto_tokens = encode_input_as_tokens(onto_test_data)

len(onto_tokens)

onto_pred = predict_ner(model, onto_tokens)

onto_pred_decoded = decode_prediction(onto_pred)

onto_target_text = onto_test_data['target_text']

onto_accuracy = compare_prediction_with_target(onto_pred_decoded, onto_target_text)
onto_accuracy

evaluations(predictions=conell_pred_decoded, references=conell_target_text):

## wnut test data 

wnut_test_data = read_test_data('./1-NER_Data/1-csv_format/test/wnut_test_data.csv')
wnut_test_data

wnut_tokens = encode_input_as_tokens(wnut_test_data)

len(wnut_tokens)

wnut_pred = predict_ner(model, wnut_tokens)

wnut_pred_decoded = decode_prediction(wnut_pred)

wnut_target_text = wnut_test_data['target_text']

wnut_accuracy = compare_prediction_with_target(wnut_pred_decoded, wnut_target_text)
wnut_accuracy

evaluations(predictions=wnut_pred_decoded, references=wnut_target_text):
