==> 5-NEL_Preprocessing_2.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

def remove_quotes(value):
    value = value.replace("['", "[")
    value = value.replace("']", "]")
    value = value.replace(", '", ", ")
    value = value.replace("',", ",")
    return value

def remove_outer_bracket(value):
    return value[1:-1]

def add_asterisks(value):
    return '**' + value + '**'

# split the data set into 70% train,  20% test, 10% valid
complete_dataset = pd.read_csv('./2-NEL_Data/1-csv_format_1/webqsp_dataset.csv')
len(complete_dataset)

train_val = complete_dataset.sample(frac=0.8,random_state=42) #random state is a seed value
web_test = complete_dataset.drop(train_val.index)

web_val = train_val.sample(frac=0.125,random_state=42)
web_train = train_val.drop(web_val.index)

web_train = web_train.reset_index(drop=True)
web_test = web_test.reset_index(drop=True)
web_val = web_val.reset_index(drop=True)

# print(len(web_train))
# print(len(web_train) / len(complete_dataset))
# print(len(web_test))
# print(len(web_test) / len(complete_dataset))
# print(len(web_val))
# print(len(web_val) / len(complete_dataset))

simple_train = pd.read_csv('./2-NEL_Data/1-csv_format_1/simple_questions_training_dataset.csv')

training_data = pd.concat([simple_train, web_train])

training_data['entity'] = training_data['entity'].apply(remove_quotes)
training_data['wikidata_reply'] = training_data['wikidata_reply'].apply(remove_quotes)

training_data['entity'] = training_data['entity'].apply(remove_outer_bracket)
training_data['wikidata_reply'] = training_data['wikidata_reply'].apply(remove_outer_bracket)

training_data.reset_index(inplace=True,drop=True)
training_data

len(training_data[training_data['wikidata_reply'].str.len() <= 2].index)

training_data = training_data.drop(training_data[training_data['wikidata_reply'].str.len() <= 2].index)
training_data.reset_index(inplace=True,drop=True)
training_data

training_data['question'] = training_data['question'].apply(add_asterisks)
training_data['entity'] = training_data['entity'].apply(add_asterisks)
training_data['wikidata_reply'] = training_data['wikidata_reply'].apply(add_asterisks)
training_data

training_data.to_csv('./2-NEL_Data/2-csv_format_2/training_data.csv', index=False)

simple_val = pd.read_csv('./2-NEL_Data/1-csv_format_1/simple_questions_val_dataset.csv')

val_data = pd.concat([simple_val, web_val])

val_data['entity'] = val_data['entity'].apply(remove_quotes)
val_data['wikidata_reply'] = val_data['wikidata_reply'].apply(remove_quotes)

val_data['entity'] = val_data['entity'].apply(remove_outer_bracket)
val_data['wikidata_reply'] = val_data['wikidata_reply'].apply(remove_outer_bracket)

val_data.reset_index(inplace=True,drop=True)
val_data

len(val_data[val_data['wikidata_reply'].str.len() <= 2].index)

val_data=val_data.drop(val_data[val_data['wikidata_reply'].str.len() <= 2].index)
val_data.reset_index(inplace=True,drop=True)
val_data

val_data['question'] = val_data['question'].apply(add_asterisks)
val_data['entity'] = val_data['entity'].apply(add_asterisks)
val_data['wikidata_reply'] = val_data['wikidata_reply'].apply(add_asterisks)
val_data

val_data.to_csv('./2-NEL_Data/2-csv_format_2/val_data.csv', index=False)

simple_test = pd.read_csv('./2-NEL_Data/1-csv_format_1/simple_questions_test_dataset.csv')

test_data = pd.concat([simple_test, web_test])

test_data['entity'] = test_data['entity'].apply(remove_quotes)
test_data['wikidata_reply'] = test_data['wikidata_reply'].apply(remove_quotes)

test_data['entity'] = test_data['entity'].apply(remove_outer_bracket)
test_data['wikidata_reply'] = test_data['wikidata_reply'].apply(remove_outer_bracket)

test_data.reset_index(inplace=True,drop=True)
test_data

len(test_data[test_data['wikidata_reply'].str.len() <= 2].index)

test_data=test_data.drop(test_data[test_data['wikidata_reply'].str.len() <= 2].index)
test_data.reset_index(inplace=True,drop=True)

test_data['question'] = test_data['question'].apply(add_asterisks)
test_data['entity'] = test_data['entity'].apply(add_asterisks)
test_data['wikidata_reply'] = test_data['wikidata_reply'].apply(add_asterisks)

test_data

test_data.to_csv('./2-NEL_Data/2-csv_format_2/test_data.csv', index=False)
