==> 7-NEL_evaluation.ipynb <==
# https://towardsdatascience.com/calculating-string-similarity-in-python-276e18a7d33a
# https://pypi.org/project/fuzzywuzzy/
# https://www.adamsmith.haus/python/answers/how-to-find-a-similarity-metric-between-two-strings-in-python


from transformers import T5TokenizerFast, T5ForConditionalGeneration 
from transformers import Trainer

import os

os.environ["CUDA_VISIBLE_DEVICES"] = "1,2,3" 
import pandas as pd
import numpy as np
import torch
import torchvision
import Levenshtein
from fuzzywuzzy import fuzz


# device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
device = torch.device("cuda:2") if torch.cuda.is_available() else torch.device("cpu")

# Create torch dataset
# https://towardsdatascience.com/fine-tuning-pretrained-nlp-models-with-huggingfaces-trainer-6326a4456e7b

class Dataset(torch.utils.data.Dataset):
    def __init__(self, encodings, labels=None):
        self.encodings = encodings
        self.labels = labels

    def __getitem__(self, idx):
        item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
        if self.labels:
            item["labels"] = torch.tensor(self.labels['input_ids'][idx])
        return item

    def __len__(self):
        return len(self.encodings["input_ids"])

def get_levenshtein_dis(str_1, str_2):
    return Levenshtein.distance(str_1, str_2)

def get_fuzzy_ration(str_1, str_2):
    return fuzz.ratio(str_1, str_2)

max_source_length = 1024
max_target_length = 128

tokenizer = T5TokenizerFast.from_pretrained("t5-base")
model_path = "./NEL_model_normal/checkpoint-5500"
model = T5ForConditionalGeneration.from_pretrained(model_path).to(device)

test_data = pd.read_csv('./2-NEL_Data/2-csv_format_2/test_data.csv')
test_data = test_data.sample(frac=1, random_state=1)
test_data

input_text = list(test_data['question'] + ',' + test_data['entity'] + ',' + test_data['wikidata_reply'])
input_text[0]

target_text = list(test_data['qid'])
target_text[0]

X_test_tokenized = tokenizer(['nel: ' + sequence for sequence in input_text], 
                              padding=True, 
                              truncation=True, 
                              max_length=max_source_length)

y_test_tokenized = tokenizer(target_text, 
                              padding=True, 
                              truncation=True, 
                              max_length=max_target_length)

print(len(test_data))

from transformers import Seq2SeqTrainingArguments

test_args = Seq2SeqTrainingArguments(
    "test_trainer",
    per_device_train_batch_size = 2,
    per_device_eval_batch_size = 2,
    eval_accumulation_steps = 50,  # VIP
    predict_with_generate=True
)

from transformers import Seq2SeqTrainer

trainer = Seq2SeqTrainer(
    model=model, 
    args=test_args
#     train_dataset= train_dataset, 
#     eval_dataset= eval_dataset
)

test_dataset = Dataset(X_test_tokenized, y_test_tokenized) 

    # Define test trainer
test_trainer = Seq2SeqTrainer(model, args=test_args, tokenizer=tokenizer)

    # Make prediction
raw_pred = test_trainer.predict(test_dataset)

predicitons = tokenizer.batch_decode(raw_pred[1], skip_special_tokens=True)
predicitons

tokens = []
for text in input_text:
    tokens.append(tokenizer('nel: ' + text, return_tensors="pt", padding=True, truncation=True, max_length=512).input_ids)

results = []
model = model.to(device)

for token in tokens:
    results.append(model.generate(token.to(device))) 

final_ouput_2 = []
for result in results:
    final_ouput_2.append(tokenizer.decode(result[0], skip_special_tokens=True))

counter = 0
for i in range(len(final_ouput_2)):
    if final_ouput_2[i] != predicitons[i]:
        print(target_text[i] + "    " + final_ouput_2[i] + "    " + predicitons[i])
        counter += 1
counter

counter = 0
for i in range(len(final_ouput_2)):
    if predicitons[i] != target_text[i]:
        counter += 1
1- counter/len(predicitons)

counter = 0
for i in range(len(final_ouput_2)):
    if final_ouput_2[i] != target_text[i]:
        counter += 1
1- counter/len(predicitons)
