==> 8-Relation_linking_preprocessing.ipynb <==
import json
import pandas as pd
import time

# refer to this paper
# https://arxiv.org/pdf/2108.07337.pdf

def get_relation_label(rel_id):
    label = ''
    API_ENDPOINT = "https://www.wikidata.org/w/api.php"
    params = {
        'action': 'wbgetentities',
        'format': 'json',
        'languages': 'en',
        'props': 'labels',
        'ids': ''
    }
      
    params['ids'] = str(rel_id)

    try:
        response = requests.get(API_ENDPOINT, params = params).json()['entities']
        label = response[str(rel_id)]['labels']['en']['value']
    except:
        return (rel_id, label)
    return (rel_id, label)

def get_entity_relation_ids(ent_id):
    relations = []
    API_ENDPOINT = "https://www.wikidata.org/w/api.php"
    params = {
        'action': 'wbgetentities',
        'format': 'json',
        'languages': 'en',
        'props': 'claims',
        'ids': ''
    }
        
    params['ids'] = str(ent_id)
    try:
        response = requests.get(API_ENDPOINT, params = params).json()['entities']
        relations = list(response[str(ent_id)]['claims'].keys())
    except:
        return (ent_id, relations)
    return (ent_id, relations)

def get_entity_labels_and_aliases(ent_id):
    labels = []
    API_ENDPOINT = "https://www.wikidata.org/w/api.php"
    params = {
        'action': 'wbgetentities',
        'format': 'json',
        'languages': 'en',
        'props': 'labels|aliases',    
        'ids': ''
    }

        
    params['ids'] = str(ent_id)
    try:
        response = requests.get(API_ENDPOINT, params = params).json()['entities']
        labels.append(response[str(ent_id)]['labels']['en']['value'])
        
        for alias in response[str(ent_id)]['aliases']['en']:
            labels.append(alias['value'])
    except:
        return (ent_id, labels)
    return (ent_id, labels)

from fuzzywuzzy import fuzz

# when creating a function the arguments are (questions and the results wikidata api after lower())
def find_nearest_word_in_question(questions, lower_wikidata_replies):
    entities_list = []
    found_match = False
    counter = 0
    for i in range(len(questions)):
        found_match = False
        for label in lower_wikidata_replies[i][1]:
            if label in questions[i].lower(): 
                found_match = True
                index = questions[i].lower().find(label)
                entities_list.append(questions[i][index : len(label)+index])
                break
        if not found_match:
            split_tokens = questions[i][1].split()
            list_of_comb = []
            list_of_comb.extend(split_tokens)
            scores = []
            for j in range(len(split_tokens)-1):
                list_of_comb.append(split_tokens[j] + ' ' + split_tokens[j+1])
            for j in range(len(split_tokens)-2):
                list_of_comb.append(split_tokens[j] + ' ' + split_tokens[j+1] + ' ' + split_tokens[j+2])
            for j in range(len(split_tokens)-3):
                list_of_comb.append(split_tokens[j] + ' ' + split_tokens[j+1] + ' ' + split_tokens[j+2] + ' ' + split_tokens[j+3])

            for label in lower_wikidata_replies[i][1]:
                for token in list_of_comb: 
                    scores.append((fuzz.ratio(label, token.lower()), token, label))
            scores = sorted(scores, key=lambda x: x[0], reverse=True)

            if len(scores) and scores[0][0] < 85:
                counter += 1
                entities_list.append('')
            elif not(len(scores)):
                counter += 1
                entities_list.append('')
            elif scores[0][0] >= 85:
                entities_list.append(scores[0][1])
    return (entities_list, counter)

# jupyter notebook never finishes processing using multiprocessing python 3
# https://stackoverflow.com/questions/47313732/jupyter-notebook-never-finishes-processing-using-multiprocessing-python-3

from multiprocessing import Pool
from functools import partial
import inspect
import time
import os
import requests


def parallal_task(func, iterable, *params): 
    with open(f'./tmp_func.py', 'w') as file:
        file.write("import requests \n")
        file.write(inspect.getsource(func).replace(func.__name__, 'task'))

    from tmp_func import task
    pool = Pool(processes=15)
    res = pool.map(task, iterable)
    pool.close()
    
    os.remove('./tmp_func.py')
    return res
#     else:
#         raise "Not in Jupyter Notebook"

from multiprocessing import Pool
from functools import partial
import inspect
import time
import os
import requests


def parallal_task_2(func, iterable, *params): 
    with open(f'./tmp_func_2.py', 'w') as file:
        file.write("import requests \n")
        file.write(inspect.getsource(func).replace(func.__name__, 'task'))

    from tmp_func_2 import task
    pool = Pool(processes=15)
    res = pool.map(task, iterable)
    pool.close()
    
    os.remove('./tmp_func_2.py')
    return res

from multiprocessing import Pool
from functools import partial
import inspect
import time
import os
import requests


def parallal_task_3(func, iterable, *params): 
    with open(f'./tmp_func_3.py', 'w') as file:
        file.write("import requests \n")
        file.write(inspect.getsource(func).replace(func.__name__, 'task'))

    from tmp_func_2 import task
    pool = Pool(processes=15)
    res = pool.map(task, iterable)
    pool.close()
    
    os.remove('./tmp_func_3.py')
    return res

## Simple_questions Training Data

with open('2-NEL_Data/0-raw_data/simple_questions_v2/annotated_wd_data_train.txt', encoding='utf-8') as file:
    lines = file.readlines()

# print(len(lines))
training_data = []
for line in lines:
    line = line.strip().split('\t', 3)
    del line[2]
    
    training_data.append(line)
    
qids = [line[0] for line in training_data]
target_relation = [line[1].replace('R', 'P') for line in training_data]
questions = [line[2] for line in training_data]
# print(len(training_data))
# print(training_data[0])
# print(qids[0])
# print(questions[0])

print(target_relation[0])
len(target_relation)

get_entity_relation_ids('Q126399')[1]

# qids[0]


target_relation[0]
# target_relation[0] in get_entity_relation_ids('Q126399')[1]

start = time.time()
relation_ids_results = parallal_task(get_entity_relation_ids, qids)
print(time.time() - start)
len(relation_ids_results)

relation_ids_results[:2]

unique_relation_ids = set()
all_test = []
for couple in relation_ids_results:
    unique_relation_ids.update(couple[1])

len(unique_relation_ids)

start = time.time()
result_relation_labels = parallal_task_2(get_relation_label, unique_relation_ids)
print(time.time() - start)

relation_dict = dict()
for (relation_id, label) in result_relation_labels:
    relation_dict[relation_id] = label

# target_relation

target_relation_labels = []
for id in target_relation:
    if id in relation_dict.keys():
        target_relation_labels.append(relation_dict[id])

len(target_relation_labels)*100/len(target_relation)

relation_strs = pd.DataFrame(target_relation_labels, columns=['label'])
relation_strs.head()

relation_strs['label'].str.len().hist(bins=30)

f'{len(relation_strs[relation_strs["label"].str.len()> 40])*100/len(relation_strs)}%'

entities_with_relation_labels = []
for (entity_id, relation_ids) in relation_ids_results:
    relation_labels = ''
    for relation_id in relation_ids:
        if len(relation_dict[relation_id]) < 40:
            relation_labels += f' {relation_dict[relation_id]} , '
    
    relation_labels = f'[{relation_labels[:-2]}]'
    entities_with_relation_labels.append([entity_id, relation_labels])

entities_with_relation_labels

get_entity_labels_and_aliases(qids[0])

start = time.time()
entity_labels_and_aliases = parallal_task_3(get_entity_labels_and_aliases, qids)
# entity_labels_and_aliases = [get_entity_labels_and_aliases(qid) for qid in qids]
print(time.time() - start)

lower_results = []
for result in entity_labels_and_aliases:
    results_for_one = [label.lower() for label in result]
    lower_results.append((result[0], results_for_one))
lower_results[:2]

(entities_list, counter) = find_nearest_word_in_question(questions, lower_results)

rows = []
for i in range(len(questions)):
    rows.append([questions[i], entities_list[i], entities_with_relation_labels[i][1], target_relation[i]])

# training_data_results = pd.DataFrame([questions[:10], entities_with_relation_labels, columns= ['entity id', 'relation labels'])
training_data_results = pd.DataFrame(rows, columns=['question', 'entity label', 'relation labels', 'target relation id'])
training_data_results

training_data_results.to_csv('./3-Relation_Linking_Data/1-csv_format/training_data.csv', index=False)

## Simple_questions Validation set

with open('2-NEL_Data/0-raw_data/simple_questions_v2/annotated_wd_data_valid.txt', encoding='utf-8') as file:
    lines = file.readlines()

# print(len(lines))
validation_data = []
for line in lines:
    line = line.strip().split('\t', 3)
    del line[2]
    validation_data.append(line)
    
qids = [line[0] for line in validation_data]
target_relation = [line[1].replace('R', 'P') for line in validation_data]
questions = [line[2] for line in validation_data]

start = time.time()
relation_ids_results = parallal_task(get_entity_relation_ids, qids)
print(time.time() - start)
len(relation_ids_results)

unique_relation_ids = set()
all_test = []
for couple in relation_ids_results:
    unique_relation_ids.update(couple[1])

start = time.time()
result_relation_labels = parallal_task_2(get_relation_label, unique_relation_ids)
print(time.time() - start)

relation_dict = dict()
for (relation_id, label) in result_relation_labels:
    relation_dict[relation_id] = label

entities_with_relation_labels = []
for (entity_id, relation_ids) in relation_ids_results:
    relation_labels = ''
    for relation_id in relation_ids:
        if len(relation_dict[relation_id]) < 40:
            relation_labels += f' {relation_dict[relation_id]} , '
    
    relation_labels = f'[{relation_labels[:-2]}]'
    entities_with_relation_labels.append([entity_id, relation_labels])

start = time.time()
entity_labels_and_aliases = parallal_task_3(get_entity_labels_and_aliases, qids)
# entity_labels_and_aliases = [get_entity_labels_and_aliases(qid) for qid in qids]
print(time.time() - start)

lower_results = []
for result in entity_labels_and_aliases:
    results_for_one = [label.lower() for label in result]
    lower_results.append((result[0], results_for_one))
lower_results[:2]

(entities_list, counter) = find_nearest_word_in_question(questions, lower_results)

rows = []
for i in range(len(questions)):
    rows.append([questions[i], entities_list[i], entities_with_relation_labels[i][1], target_relation[i]])

validation_data_results = pd.DataFrame(rows, columns=['question', 'entity label', 'relation labels', 'target relation id'])
validation_data_results.head(10)

validation_data_results.to_csv('./3-Relation_Linking_Data/1-csv_format/validation_data.csv', index=False)

## Simple_questions Testing Data

with open('2-NEL_Data/0-raw_data/simple_questions_v2/annotated_wd_data_test.txt', encoding='utf-8') as file:
    lines = file.readlines()

# print(len(lines))
test_data = []
for line in lines:
    line = line.strip().split('\t', 3)
    del line[2]
    test_data.append(line)

    
qids = [line[0] for line in test_data]
target_relation = [line[1].replace('R', 'P') for line in test_data]
questions = [line[2] for line in test_data]

len(qids)

start = time.time()
relation_ids_results = parallal_task(get_entity_relation_ids, qids)
print(time.time() - start)
len(relation_ids_results)

# relation_ids_results

unique_relation_ids = set()
all_test = []
for couple in relation_ids_results:
    unique_relation_ids.update(couple[1])

start = time.time()
result_relation_labels = parallal_task_2(get_relation_label, unique_relation_ids)
print(time.time() - start)

relation_dict = dict()
for (relation_id, label) in result_relation_labels:
    relation_dict[relation_id] = label

target_relation

# target_relation_labels = []
# for id in target_relation:
#     if id in relation_dict.keys():
#         target_relation_labels.append(relation_dict[id])

# len(target_relation_labels)

entities_with_relation_labels = []
for (entity_id, relation_ids) in relation_ids_results:
    relation_labels = ''
    for relation_id in relation_ids:
        if len(relation_dict[relation_id]) < 40:
            relation_labels += f' {relation_dict[relation_id]} , '
    
    relation_labels = f'[{relation_labels[:-2]}]'
    entities_with_relation_labels.append([entity_id, relation_labels])

start = time.time()
entity_labels_and_aliases = parallal_task_3(get_entity_labels_and_aliases, qids)
# entity_labels_and_aliases = [get_entity_labels_and_aliases(qid) for qid in qids]
print(time.time() - start)

lower_results = []
for result in entity_labels_and_aliases:
    results_for_one = [label.lower() for label in result]
    lower_results.append((result[0], results_for_one))
lower_results[:2]

(entities_list, counter) = find_nearest_word_in_question(questions, lower_results)

c_empty = 0
for e in entities_list:
    if e == '':
        c_empty += 1
print(c_empty)
print(c_empty*100/len(entities_list))

len(questions)

target_relation_labels = []
for id in target_relation:
    if id in relation_dict.keys():
        target_relation_labels.append(relation_dict[id])
    else:
        print(id)
        target_relation_labels.append('')

get_relation_label('P738')

len(target_relation_labels)

rows = []
for i in range(len(questions)):
    rows.append([questions[i], entities_list[i], entities_with_relation_labels[i][1], target_relation[i], target_relation_labels[i]])

test_data_results = pd.DataFrame(rows, columns=['question', 'entity label', 'relation labels', 'target relation id', 'target relation labels'])
test_data_results.head(10)

test_data_results

test_data_results.to_csv('./3-Relation_Linking_Data/1-csv_format/test_data.csv', index=False)
