==> 90-Relation_linking_final_preprocessing.ipynb <==
import pandas as pd

def clean_row_ids(row):
    row['relation_id'] = row['relation_id'].replace('dict_keys(', '')
    row['relation_id'] = row['relation_id'].replace(')', '')
    row['relation_id'] = row['relation_id'].replace("'", '')
    row['relation_id'] = row['relation_id'].replace("[", '')
    row['relation_id'] = row['relation_id'].replace("]", '')

    return row['relation_id']

def clean_relation_labels(row):
    row['relation_labels'] = row['relation_labels'].replace('dict_values(', '')
    row['relation_labels'] = row['relation_labels'].replace(')', '')
    row['relation_labels'] = row['relation_labels'].replace("'", '')
    row['relation_labels'] = row['relation_labels'].replace("[", '')
    row['relation_labels'] = row['relation_labels'].replace("]", '')
    
    return row['relation_labels']

def clean_relation_labels_original(row):
    row['relation labels'] = row['relation labels'].replace('dict_values(', '')
    row['relation labels'] = row['relation labels'].replace(')', '')
    row['relation labels'] = row['relation labels'].replace("'", '')
    row['relation labels'] = row['relation labels'].replace("[", '')
    row['relation labels'] = row['relation labels'].replace("]", '')
    
    return row['relation labels']

def combine_relations(row):
    relations = []
    relations_len = 0
    
    relations_obj = str(row['relation_obj_labels']).split(', ')
    if len(relations_obj) == 0:
        return row['relation labels']

    relations_sub = str(row['relation labels']).split(' , ')

    for i in range(max(len(relations_sub), len(relations_obj))):
        if i < len(relations_sub) and relations_sub[i] not in relations:
            relations.append(relations_sub[i].strip())
            
        if i < len(relations_obj) and relations_obj[i].strip() and relations_obj[i] not in relations and len(relations_obj[i]) < 40:
            relations.append(relations_obj[i].strip())
                   
    return relations

## Training Data

training_data_obj_1 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/training_data_find_obj.csv')
training_data_obj_2 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/training_data_find_obj_2.csv')

training_data_obj = pd.concat([training_data_obj_1, training_data_obj_2])

training_data_obj['relation_id'] = training_data_obj.apply(clean_row_ids, axis=1)
training_data_obj['relation_labels'] = training_data_obj.apply(clean_relation_labels, axis=1)

training_data_obj.head()

training_data_obj = training_data_obj.drop_duplicates(subset=['questions', 'target_relation'])

len(training_data_obj)

training_data_obj.head()

training_data_obj = training_data_obj.rename(columns={'questions': 'question', 'relation_id': 'relation_obj_labels', 'relation_labels': 'relation_obj_ids', 'target_relation': 'target relation id'})

training_data_obj 

training_data = pd.read_csv('./3-Relation_Linking_Data/1-csv_format/training_data.csv').drop_duplicates()
len(training_data)

training_data['relation labels'] = training_data.apply(clean_relation_labels_original, axis= 1)

training_data

new_training_data = pd.merge(training_data, training_data_obj, how="left", on='question')

new_training_data = new_training_data.drop_duplicates(subset=['question', 'target relation id_x', 'target relation id_y'])

new_training_data = new_training_data.reset_index(drop=True)

new_training_data

new_training_data['combined relations'] = new_training_data.apply(combine_relations, axis = 1)

new_training_data

new_training_data.to_csv('./3-Relation_Linking_Data/2-final_input/training_data.csv', index=False)

## Validation Data

val_data_obj_1 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/val_data_find_obj.csv')
val_data_obj_2 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/val_data_find_obj_2.csv')

val_data_obj = pd.concat([val_data_obj_1, val_data_obj_2])

val_data_obj['relation_id'] = val_data_obj.apply(clean_row_ids, axis=1)
val_data_obj['relation_labels'] = val_data_obj.apply(clean_relation_labels, axis=1)

val_data_obj = val_data_obj.drop_duplicates(subset=['questions', 'target_relation'])

val_data_obj = val_data_obj.rename(columns={'questions': 'question', 'relation_id': 'relation_obj_labels', 'relation_labels': 'relation_obj_ids', 'target_relation': 'target relation id'})

val_data = pd.read_csv('./3-Relation_Linking_Data/1-csv_format/validation_data.csv').drop_duplicates()
len(val_data)

val_data['relation labels'] = val_data.apply(clean_relation_labels_original, axis= 1)

new_val_data = pd.merge(val_data, val_data_obj, how="left", on='question')

new_val_data = new_val_data.drop_duplicates(subset=['question', 'target relation id_x', 'target relation id_y'])

new_val_data = new_val_data.reset_index(drop=True)

new_val_data['combined relations'] = new_val_data.apply(combine_relations, axis = 1)

new_val_data.to_csv('./3-Relation_Linking_Data/2-final_input/validation_data.csv', index=False)

new_val_data

len(new_val_data)

## Test Data

test_data_obj_1 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/test_data_find_obj.csv')
test_data_obj_2 = pd.read_csv('./3-Relation_Linking_Data/0-preprocess/test_data_find_obj_2.csv')

test_data_obj = pd.concat([test_data_obj_1, test_data_obj_2])

test_data_obj['relation_id'] = test_data_obj.apply(clean_row_ids, axis=1)
test_data_obj['relation_labels'] = test_data_obj.apply(clean_relation_labels, axis=1)

test_data_obj = test_data_obj.drop_duplicates(subset=['questions', 'target_relation'])

test_data_obj = test_data_obj.rename(columns={'questions': 'question', 'relation_id': 'relation_obj_labels', 'relation_labels': 'relation_obj_ids', 'target_relation': 'target relation id'})

test_data = pd.read_csv('./3-Relation_Linking_Data/1-csv_format/test_data.csv').drop_duplicates()
len(test_data)

test_data.head()

test_data_obj.head()

test_data['relation labels'] = test_data.apply(clean_relation_labels_original, axis= 1)

test_data

new_test_data = pd.merge(test_data, test_data_obj, how="left", on='question')

new_test_data.columns

new_test_data = new_test_data.drop_duplicates(subset=['question', 'target relation id_x', 'target relation id_y'])

new_test_data

new_test_data = new_test_data.reset_index(drop=True)

new_test_data['combined relations'] = new_test_data.apply(combine_relations, axis = 1)

new_test_data.to_csv('./3-Relation_Linking_Data/2-final_input/test_data.csv', index=False)

len(new_test_data)
